pyproject.toml
[project]
name = "day-sheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ctk>=0.1",
    "isort>=7.0.0",
    "notify-py>=0.3.43",
    "pyqt5>=5.15.11",
    "pyqt6>=6.10.1",
    "python-daemon>=3.1.2",
    "schedule>=1.2.2",
]
